district_level_pureml/requirements.txt
absl-py==2.1.0
affine==2.4.0
annotated-types==0.6.0
appdirs==1.4.4
astunparse==1.6.3
attrs==22.2.0
blinker==1.7.0
blis==0.7.11
brotlipy==0.7.0
cachetools==5.3.3
catalogue==2.0.10
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
data==0.4
docker-pycreds==0.4.0
et-xmlfile==1.1.0
filelock==3.13.4
Fiona==1.9.2
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.25.0
fsspec==2024.3.1
funcsigs==1.0.2
future==0.18.3
gast==0.4.0
gensim==4.3.2
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.22.2
intel-openmp==2021.4.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
keras==2.13.1
langcodes==3.3.0
latex==0.7.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
mkl==2021.4.0
mkl-fft==1.3.1
mkl-service==2.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
munch==2.5.0
munkres==1.1.4
murmurhash==1.0.10
namex==0.0.7
node2vec==0.4.6
numpy==1.24.3
oauthlib==3.2.2
openpyxl==3.1.4
opt-einsum==3.3.0
optree==0.11.0
pathtools==0.1.2
Pillow==9.2.0
ply==3.11
preshed==3.0.9
protobuf==4.22.1
psutil==5.9.4
PuLP==2.8.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
pyg-lib==0.4.0+pt23cu121
pygame==2.4.0
pyglet==1.5.27
Pyomo==6.7.1
PyPDF2==3.0.1
pyproj==3.5.0
PyQt5==5.12.3
PyQt5_sip==4.19.18
PyQtChart==5.12
PyQtWebEngine==5.12.1
pywin32==305.1
PyYAML==6.0
rasterio==1.3.6
rdflib==7.0.0
regex==2024.4.16
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.3.0
scipy==1.13.1
sentry-sdk==1.18.0
setproctitle==1.3.2
shapely==2.0.1
shutilwhich==1.1.0
smart-open==6.4.0
smmap==5.0.0
snuggs==1.4.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tbb==2021.11.0
tempdir==0.7.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.19.1
torch==2.3.1+cu121
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchtext==0.17.2
tqdm==4.64.1
transformers==4.40.1
typer==0.9.4
typing_extensions==4.5.0
wandb==0.14.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wincertstore==0.2
wrapt==1.16.0
